services/image_gen/requirements.txt
fastapi
uvicorn[standard]
diffusers>=0.27.2
transformers
accelerate
safetensors
Pillow
